requirements.txt
elasticsearch==8.1.1
Flask==2.1.0
PyJWT==2.3.0


